[R] Fix windows development build
The `configure.win` script copies the c++ files into the R package because Windows does not understand the symlinks. This makes it possible to install the source package on windows straight from git.

Note that this is not needed for released versions of feather because `R CMD build` already replaces symlinks with actual content in the source package. Hence `configure.win` in `.Rbuildignore`.

Author: Jeroen Ooms <jeroenooms@gmail.com>

Closes #201 from jeroenooms/windows and squashes the following commits:

30e7497 [Jeroen Ooms] Fix windows development build