[R/Python] Metadata handling tweaks, implement `columns` filter argument in Python
Both the Python and R code were using `TableReader::GetColumn` to access column metadata which causes data to be read from the file. This is OK when we are using memory maps but will cause performance problems should we switch to some other file-like object. I added a `GetColumnMetadata` API and refactored both Python and R codebases to use this

Closes #208

Author: roman-kh <rhudor@gmail.com>
Author: Wes McKinney <wesm@apache.org>

Closes #218 from wesm/feature/python-read-columns and squashes the following commits:

8deeaf8 [Wes McKinney] Expose column metadata in public API to avoid reading columns multiple times. Fix up Python and R code
b5f2b68 [roman-kh] [Python] Add support of columns parameter in read_dataframe