Workaround for Error Messages
Under certain conditions, Rscript gives a warning message of:

    WARNING: ignoring environment value of R_HOME

When there seems to be a (valid but unusual) setup of R_HOME.  As a
result, the command to determine the endianess of the system returns

    WARNING: ignoring environment value of R_HOME little

instead of just “little”. causing this script to think it is little
endian when it isn’t.  This pull runs the command wrapped in a
suppressWarnings call so that only the platform endianess and not the
warning are returned.