[JS] Add support for JS
This PR partially addresses issue #17. There's plenty of missing features, but if this style of API seems OK for you, then maybe it makes sense to add support for those.

Let me know if you have concerns about documentation. There isn't a test suite, but there is a self-contained example in the test directory which, if the code works, should recreate the most boring scatterplot on the planet (I blame @hadley for including iris.feather as an example).

Author: Carlos Scheidegger <cscheid@cs.arizona.edu>

Closes #233 from cscheid/feather-js-proposal and squashes the following commits:

978668e [Carlos Scheidegger] Clean up tilde files
60a3771 [Carlos Scheidegger] Remove accidentally-commited file
93a0407 [Carlos Scheidegger] Clean up tilde files
0321a5b [Carlos Scheidegger] Add a minimal JS API and a working example