[R] Workaround for Error Messages
Under certain conditions, Rscript gives a warning message of:

    WARNING: ignoring environment value of R_HOME

When there seems to be a valid but unusual setup of R_HOME.  As a
result, the command to determine the endianess of the system returns

    WARNING: ignoring environment value of R_HOME little

instead of just “little”. causing this script to think it is big
endian when it isn’t.  This pull changes the configure script to
remove this warning message before checking the endianness.

Author: Nigel Delaney <spamavoid@outlook.com>

Closes #237 from evolvedmicrobe/warningfix and squashes the following commits:

bfb870d [Nigel Delaney] Workaround for Error Messages